_file_io/_helpers: consolidate __name__ magic in async_wraps
This also adds signature tests for AsyncIO and AsyncPath.
